refactor: use own request router instead of implicit DefaultServeMux.
Quote from https://www.alexedwards.net/blog/a-recap-of-request-handling:
%
Generally you shouldn't use the DefaultServeMux because it poses a security risk.
Because the DefaultServeMux is stored in a global variable, any package is able to access it and
register a route – including any third-party packages that your application imports. If one of those
third-party packages is compromised, they could use the DefaultServeMux to expose a malicious
handler to the web.
%